set clipCircle precision to 2 degrees (#282)
* default geoCircle precision to 2 degrees

Results in 180 control points instead of 60 (⨉3). The error is quadratic with the precision, with the worst case (for a parallel) being at latitude 45°. Under this new setting, the maximum error is now ~484m, vs. 4.365km with precision 6°.
https://observablehq.com/@d3/geocircle-error

* document

* refine clipCircle precision, from 6 degrees to 2 degrees
following https://github.com/d3/d3-geo/pull/281

(note that for clipCircle the precision is not modifiable by the user)

Manually rechecked each of the tests that are (slightly) modified as a consequence

---------

Co-authored-by: Mike Bostock <mbostock@gmail.com>